requirements.txt
faiss-cpu
tensorflow
keras
onnxruntime
openvino
opencv-python
mtcnn
ultralytics
numpy
streamlit
streamlit-webrtc
Pillow